Explain ignoring SSL verification in OSX in AWS
It would seem that for OSX instances in AWS that SSL certificates don't
have the certificat authorities setup correctly or at all. Which means
when access any website using SSL with ansible then the SSL verification
fails. This notes on how to tell ansible at the top level to not try and
validate certificates.